Created and updated file location and imports for test_decorators.py (previously test_user.py) for consistency and in response to user.py test case fail.
